+++ src/olorenchemengine/beta/__init__.py
@@ -1,0 +1,1 @@
+from .gnn import *
